testing on task 1 ,using pip3 with single quotes, and added only if exploitiong pgrep to check if there is a process called killmenow before attempt to terminated it in task 2

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,6 @@
 # create a manifest that kills a process named "killmenow".
 exec { 'kill_killmenow':
   command => 'pkill  -xf \'/bin/bash ./killmenow\'',
+  onlyif  => 'pgrep -af \'/bin/bash ./killmenow\'',
   path    => '/usr/bin/',
 }
